+++ src/libretro/screenlayout.hpp
@@ -318,7 +318,13 @@ namespace MelonDsDs {
 
     constexpr unsigned MaxSoftwareRenderedHeight() noexcept {
         using namespace config::screen;
-        return NDS_SCREEN_HEIGHT * 2 + MAX_SCREEN_GAP;
+        return std::max({
+            // Top/Bottom or Bottom/Top layout
+            NDS_SCREEN_HEIGHT * 2 + MAX_SCREEN_GAP,
+
+            // Hybrid layout
+            NDS_SCREEN_HEIGHT * MAX_HYBRID_RATIO,
+        });
     }
 
     constexpr unsigned MaxOpenGlRenderedWidth() noexcept {