+++ test/cmake/Screen.cmake
@@ -1,0 +1,7 @@
+add_python_test(
+    NAME "Core supports hybrid sceen ratio of 3:1"
+    TEST_MODULE basics.core_run_frame
+    CONTENT "${NDS_ROM}"
+    CORE_OPTION melonds_hybrid_ratio=3
+    CORE_OPTION melonds_screen_layout1=hybrid-top
+)