+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0),
 and this project roughly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.4] - 2024-07-08
+
+### Changed
+
+- Updated the "DSi NAND Path" core option description to clarify the role of the no$gba footer.
+
+### Fixed
+
+- Fixed some log entries not being output with a newline.
+- Fixed a crash when using a hybrid screen layout with a screen ratio of 3:1.
+- Fixed DSi NAND images not being recognized if they lacked a no$gba footer
+  despite having equivalent data at offset 0xFF800. [#195](https://github.com/JesseTG/melonds-ds/issues/195)
+- Fixed the screen being rendered when using the OpenGL renderer while the emulated lid is closed,
+  which caused flickering in some games. [#214](https://github.com/JesseTG/melonds-ds/issues/214)
+
 ## [1.1.3] - 2024-06-14
 
 ### Fixed